- Added enablePublicationCheck to PrivateDataReminderSettings - Added checkbox to enable publication check in the private data reminder settings form - Added "Enable Publication Check" checkbox on sendPrivateDataRemindersForm.jsp that can override the value in the saved settings. - Added DismissPubMedSuggestionAction - Added CheckPubMedForDatasetAction to perform publication check on one dataset - New method added to PanoramaPublicNotification for sending message about published paper match - Updated PrivateDataReminderJob to check for publications

+++ panoramapublic/src/org/labkey/panoramapublic/PanoramaPublicNotification.java
@@ -361,16 +361,20 @@ public class PanoramaPublicNotification
 
     public static void postPrivateDataReminderMessage(@NotNull Journal journal, @NotNull JournalSubmission js, @NotNull ExperimentAnnotations expAnnotations,
                                                       @NotNull User submitter, @NotNull User messagePoster, List<User> notifyUsers,
-                                                      @NotNull Announcement announcement, @NotNull Container announcementsContainer, @NotNull User journalAdmin)
+                                                      @NotNull Announcement announcement, @NotNull Container announcementsContainer, @NotNull User journalAdmin,
+                                                      boolean publicationFound, @Nullable String pubMedId)
     {
-        String message = getDataStatusReminderMessage(expAnnotations, submitter, js, announcement, announcementsContainer, journalAdmin);
-        String title = "Action Required: Status Update for Your Private Data on Panorama Public";
+        String message = getDataStatusReminderMessage(expAnnotations, submitter, js, announcement, announcementsContainer, journalAdmin, publicationFound, pubMedId);
+        String title = publicationFound
+                ? "Congratulations! We Found a Publication Associated with Your Data on Panorama Public"
+                : "Action Required: Status Update for Your Private Data on Panorama Public";
         postNotificationFullTitle(journal, js.getJournalExperiment(), message, messagePoster, title, AnnouncementService.StatusOption.Closed, notifyUsers);
     }
 
     public static String getDataStatusReminderMessage(@NotNull ExperimentAnnotations exptAnnotations, @NotNull User submitter,
                                                       @NotNull JournalSubmission js,@NotNull Announcement announcement,
-                                                      @NotNull Container announcementContainer, @NotNull User journalAdmin)
+                                                      @NotNull Container announcementContainer, @NotNull User journalAdmin,
+                                                      boolean publicationFound, @Nullable String pubMedId)
     {
         String shortUrl = exptAnnotations.getShortUrl().renderShortURL();
         String makePublicLink = PanoramaPublicController.getMakePublicUrl(exptAnnotations.getId(), exptAnnotations.getContainer()).getURIString();
@@ -386,32 +390,60 @@ public class PanoramaPublicNotification
         ActionURL requestExtensionUrl = new ActionURL(PanoramaPublicController.RequestExtensionAction.class, exptAnnotations.getContainer())
                 .addParameter("shortUrlEntityId", shortUrlEntityId);
 
-        ActionURL requesDeletionUrl = new ActionURL(PanoramaPublicController.RequestDeletionAction.class, exptAnnotations.getContainer())
+        ActionURL requestDeletionUrl = new ActionURL(PanoramaPublicController.RequestDeletionAction.class, exptAnnotations.getContainer())
                 .addParameter("shortUrlEntityId",shortUrlEntityId);
 
+        ActionURL dismissPubMedUrl = new ActionURL(PanoramaPublicController.DismissPubMedSuggestionAction.class, exptAnnotations.getContainer())
+                .addParameter("shortUrlEntityId", shortUrlEntityId);
 
         ExperimentAnnotations sourceExperiment = ExperimentAnnotationsManager.get(exptAnnotations.getSourceExperimentId());
 
         StringBuilder message = new StringBuilder();
-        message.append("Dear ").append(getUserName(submitter)).append(",").append(NL2)
-                .append("We are reaching out regarding your data on Panorama Public (").append(shortUrl).append("), which has been private since ")
-                .append(dateString).append(".")
-                .append(NL2).append(bold("Title:")).append(" ").append(escape(exptAnnotations.getTitle()))
-                .append(NL2).append(bold("Is the paper associated with this work already published?"))
-                .append(NL).append("- If yes: Please make your data public by clicking the \"Make Public\" button in your folder or by clicking this link: ")
-                .append(bold(link("Make Data Public", makePublicLink)))
-                .append(". This helps ensure that your valuable research is easily accessible to the community.")
-                .append(NL).append("- If not: You have a couple of options:")
-                .append(NL).append("  - ").append(bold("Request an Extension")).append(" - If your paper is still under review, or you need additional time, please let us know by clicking ")
-                .append(bold(link("Request Extension", requestExtensionUrl.getURIString()))).append(".")
-                .append(NL).append("  - ").append(bold("Delete from Panorama Public")).append(" - If you no longer wish to host your data on Panorama Public, please click ")
-                .append(bold(link("Request Deletion", requesDeletionUrl.getURIString()))).append(". ")
-                .append("We will remove your data from Panorama Public.");
-        if (sourceExperiment != null)
+        message.append("Dear ").append(getUserName(submitter)).append(",").append(NL2);
+
+        if (publicationFound && !StringUtils.isBlank(pubMedId))
         {
-            message.append(" However, your source folder (")
-                    .append(getContainerLink(sourceExperiment.getContainer()))
-                    .append(") will remain intact, allowing you to resubmit your data in the future if you wish.");
+            // Message variant when a publication was found
+            String pubMedUrl = "https://pubmed.ncbi.nlm.nih.gov/" + pubMedId;
+            message.append("Great news! We found a publication that appears to be associated with your data on Panorama Public (")
+                    .append(shortUrl).append("), which has been private since ").append(dateString).append(".")
+                    .append(NL2).append(bold("Title:")).append(" ").append(escape(exptAnnotations.getTitle()))
+                    .append(NL2).append(bold("Publication Found:")).append(" ").append(link("PubMed ID " + pubMedId, pubMedUrl))
+                    .append(NL2).append("Since your work has been published, we encourage you to make your data public so the research community can access it alongside your publication. ")
+                    .append("You can do this by clicking the \"Make Public\" button in your data folder or by clicking this link: ")
+                    .append(bold(link("Make Data Public", makePublicLink))).append(".")
+                    .append(NL2).append(bold("If this publication is not associated with your data:"))
+                    .append(NL).append("Please let us know by ").append(bold(link("clicking here", dismissPubMedUrl.getURIString())))
+                    .append(", and we will stop suggesting this publication for your dataset.")
+                    .append(NL2).append(bold("Not ready to make your data public yet?"))
+                    .append(NL).append("If you need more time, you can:")
+                    .append(NL).append("  - ").append(bold(link("Request an Extension", requestExtensionUrl.getURIString())))
+                    .append(" - This will give you additional time before we send another reminder.")
+                    .append(NL).append("  - ").append(bold(link("Request Deletion", requestDeletionUrl.getURIString())))
+                    .append(" - If you no longer wish to host your data on Panorama Public.");
+        }
+        else
+        {
+            // Original message variant when no publication was found
+            message.append("We are reaching out regarding your data on Panorama Public (").append(shortUrl).append("), which has been private since ")
+                    .append(dateString).append(".")
+                    .append(NL2).append(bold("Title:")).append(" ").append(escape(exptAnnotations.getTitle()))
+                    .append(NL2).append(bold("Is the paper associated with this work already published?"))
+                    .append(NL).append("- If yes: Please make your data public by clicking the \"Make Public\" button in your folder or by clicking this link: ")
+                    .append(bold(link("Make Data Public", makePublicLink)))
+                    .append(". This helps ensure that your valuable research is easily accessible to the community.")
+                    .append(NL).append("- If not: You have a couple of options:")
+                    .append(NL).append("  - ").append(bold("Request an Extension")).append(" - If your paper is still under review, or you need additional time, please let us know by clicking ")
+                    .append(bold(link("Request Extension", requestExtensionUrl.getURIString()))).append(".")
+                    .append(NL).append("  - ").append(bold("Delete from Panorama Public")).append(" - If you no longer wish to host your data on Panorama Public, please click ")
+                    .append(bold(link("Request Deletion", requestDeletionUrl.getURIString()))).append(". ")
+                    .append("We will remove your data from Panorama Public.");
+            if (sourceExperiment != null)
+            {
+                message.append(" However, your source folder (")
+                        .append(getContainerLink(sourceExperiment.getContainer()))
+                        .append(") will remain intact, allowing you to resubmit your data in the future if you wish.");
+            }
         }
 
         message.append(NL2).append("If you have any questions or need further assistance, please do not hesitate to respond to this message by ")
